+++ Samples~/Scripts/IssueAndTesting/Issue/Issue27.cs
@@ -44,5 +44,6 @@ namespace SaintsField.Samples.Scripts.IssueAndTesting.Issue
 
         [SerializeField] RCValue[] rCValues;
 
+        public UnityEvent<bool>[] onBoolValueFetched;
     }
 }